Implement MySomethingList for social media, use list item slot
